+++ packages/flutter_nucleus/lib/src/hooks.dart
@@ -2,12 +2,20 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart' hide Store;
 import 'package:flutter_nucleus/flutter_nucleus.dart';
 
-A useAtom<A>(Atom<A> atom) => use(_AtomHook(atom));
+A useAtom<A>(
+  Atom<A> atom, {
+  bool listen = true,
+}) =>
+    use(_AtomHook(atom, listen: listen));
 
 class _AtomHook<A> extends Hook<A> {
-  const _AtomHook(this.atom);
+  const _AtomHook(
+    this.atom, {
+    required this.listen,
+  });
 
   final Atom<A> atom;
+  final bool listen;
 
   @override
   _AtomHookState<A> createState() => _AtomHookState();
@@ -20,12 +28,17 @@ class _AtomHookState<A> extends HookState<A, _AtomHook<A>> {
   void _setup(BuildContext context) {
     final store = AtomScope.storeOf(context);
     _value = store.read(hook.atom);
-    _cancel = store.subscribe(
-      hook.atom,
-      () => setState(() {
-        _value = store.read(hook.atom);
-      }),
-    );
+
+    if (hook.listen) {
+      _cancel = store.subscribe(
+        hook.atom,
+        () => setState(() {
+          _value = store.read(hook.atom);
+        }),
+      );
+    } else {
+      _cancel = store.mount(hook.atom);
+    }
   }
 
   @override