+++ packages/nucleus/lib/src/atoms.dart
@@ -1,6 +1,5 @@
 export 'atoms/atom.dart';
-export 'atoms/future.dart';
+export 'atoms/parent.dart';
 export 'atoms/proxy.dart';
 export 'atoms/read_only.dart';
 export 'atoms/state.dart';
-export 'atoms/stream.dart';