+++ packages/nucleus/lib/src/atoms/read_only.dart
@@ -1,5 +1,6 @@
 import 'package:nucleus/nucleus.dart';
 
+/// See [atom].
 class ReadOnlyAtom<Value> extends Atom<Value> {
   ReadOnlyAtom(this._reader);
 
@@ -9,6 +10,11 @@ class ReadOnlyAtom<Value> extends Atom<Value> {
   Value read(_) => _reader(_);
 }
 
-// Function API
-
+/// Create a read only atom that can interact with other atom's to create
+/// derived state.
+///
+/// ```dart
+/// final count = stateAtom(0);
+/// final countTimesTwo = atom((get) => get(count) * 2);
+/// ```
 Atom<Value> atom<Value>(AtomReader<Value> create) => ReadOnlyAtom(create);