+++ packages/nucleus/lib/src/store.dart
@@ -26,7 +26,7 @@ class Store {
   var _pendingWrites = HashMap<Atom, AtomState?>();
 
   Future<void>? _schedulerFuture;
-  final _atomsScheduledForRemoval = <Atom>[];
+  final _atomsScheduledForRemoval = HashSet<Atom>();
 
   // === Public api
   Value read<Value>(Atom<Value> atom) => _read(atom).value as Value;
@@ -47,12 +47,12 @@ class Store {
   }
 
   Future<R> use<R>(Atom atom, FutureOr<R> Function() f) async {
-    final mount = _ensureMounted(atom);
+    final unmount = mount(atom);
 
     try {
       return await f();
     } finally {
-      _maybeUnmount(mount);
+      unmount();
     }
   }
 
@@ -67,7 +67,7 @@ class Store {
 
   void _setState(Atom atom, AtomState state, AtomState? previousState) {
     _atomStateMap[atom] = state;
-    _pendingWrites.putIfAbsent(atom, () => previousState);
+    _pendingWrites[atom] ??= previousState;
     _maybeScheduleAtomRemoval(atom, state);
   }
 
@@ -155,9 +155,9 @@ class Store {
     }
 
     // Needs recomputation
-    currentState?.onDispose();
+    currentState?.context.dispose();
 
-    final ctx = _ReadContext(this, atom, currentState?.value);
+    final ctx = ReadContext(this, atom, currentState?.value);
     final value = atom.$read(ctx);
 
     return ctx.calledSetSelf
@@ -190,7 +190,7 @@ class Store {
   AtomState _put(
     Atom atom,
     Object? value, {
-    _ReadContext? context,
+    ReadContext? context,
   }) {
     final currentState = _atomStateMap[atom];
 
@@ -212,8 +212,8 @@ class Store {
       revision: revision,
       valid: true,
       dependencies: deps,
-      disposers: context?.disposers,
-      keepAliveOverride: atom.keepAliveOverride,
+      context:
+          context ?? currentState?.context ?? ReadContext(this, atom, null),
     );
 
     if (currentState == nextState) {
@@ -237,6 +237,7 @@ class Store {
 
       final state = _atomStateMap[dep];
       if (state != null) {
+        state.context.dispose();
         _setState(dep, state.copyWith(valid: false), state);
       }
 
@@ -249,29 +250,22 @@ class Store {
     while (_pendingWrites.isNotEmpty) {
       final pending = _pendingWrites;
       _pendingWrites = HashMap();
-
-      for (final e in pending.entries) {
-        final atom = e.key;
-        final previousState = e.value;
-        final currentState = _atomStateMap[atom];
-
-        if (currentState != null &&
-            currentState.value != previousState?.value) {
-          _mountDependencies(atom, currentState, previousState?.dependencies);
-        }
-
-        if (previousState?.valid == false && currentState?.valid == true) {
-          continue;
-        }
-
-        final mount = _atomMountedMap[atom];
-        if (mount != null) {
-          for (final fn in mount.listeners) {
-            fn();
-          }
-        }
-      }
+      pending.forEach(_flushPendingWrite);
     }
+  }
+
+  void _flushPendingWrite(Atom atom, dynamic previousState) {
+    final currentState = _atomStateMap[atom];
+
+    if (currentState != null && currentState.value != previousState?.value) {
+      _mountDependencies(atom, currentState, previousState?.dependencies);
+    }
+
+    if (previousState?.valid == false && currentState?.valid == true) {
+      return;
+    }
+
+    _atomMountedMap[atom]?.notifyListeners();
   }
 
   void _mountDependencies(
@@ -315,10 +309,10 @@ class Store {
   void _runScheduledTasks() {
     _schedulerFuture = null;
 
-    for (final atomKey in _atomsScheduledForRemoval) {
-      if (_atomMountedMap.containsKey(atomKey)) continue;
-      _atomStateMap[atomKey]?.onDispose();
-      _atomStateMap.remove(atomKey);
+    for (final atom in _atomsScheduledForRemoval) {
+      if (_atomMountedMap.containsKey(atom)) continue;
+      _atomStateMap[atom]?.context.dispose();
+      _atomStateMap.remove(atom);
     }
     _atomsScheduledForRemoval.clear();
   }
@@ -328,7 +322,7 @@ class Store {
     AtomState state, [
     bool skipMountCheck = false,
   ]) {
-    if (state.keepAlive ||
+    if (atom.shouldKeepAlive ||
         (!skipMountCheck && _atomMountedMap.containsKey(atom))) {
       return;
     }
@@ -344,24 +338,14 @@ class AtomState {
     required this.revision,
     required this.valid,
     required this.dependencies,
-    required this.keepAliveOverride,
-    List<void Function()>? disposers,
-  }) : disposers = disposers ?? [];
+    required this.context,
+  });
 
   final Object? value;
   final int revision;
   final bool valid;
   final HashMap<Atom, int> dependencies;
-  final List<void Function()> disposers;
-
-  final bool? keepAliveOverride;
-  late final keepAlive = keepAliveOverride ?? disposers.isEmpty;
-
-  void onDispose() {
-    for (final fn in disposers) {
-      fn();
-    }
-  }
+  ReadContext context;
 
   bool hasNoDependenciesExcept(Atom atom) =>
       dependencies.isEmpty ||
@@ -376,14 +360,14 @@ class AtomState {
     int? revision,
     bool? valid,
     HashMap<Atom, int>? dependencies,
+    ReadContext? context,
   }) =>
       AtomState(
         value: value ?? this.value,
         revision: revision ?? this.revision,
         valid: valid ?? this.valid,
         dependencies: dependencies ?? this.dependencies,
-        disposers: disposers,
-        keepAliveOverride: keepAliveOverride,
+        context: context ?? this.context,
       );
 
   static const _dependencyEquality = MapEquality<Atom, int>();
@@ -413,19 +397,26 @@ class AtomMount {
   final listeners = <void Function()>[];
   final dependants = HashSet<Atom>();
 
+  void notifyListeners() {
+    for (final fn in listeners) {
+      fn();
+    }
+  }
+
   bool get isUnmountable =>
       listeners.isEmpty &&
       (dependants.isEmpty ||
           (dependants.length == 1 && dependants.contains(atom)));
 }
 
-class _ReadContext implements AtomContextBase {
-  _ReadContext(this.store, this.atom, this.previousValue);
+class ReadContext implements AtomContext<dynamic> {
+  ReadContext(this.store, this.atom, this.previousValue);
 
   final Store store;
   final Atom atom;
   final deps = HashSet<Atom>();
   final disposers = <void Function()>[];
+  var _disposed = false;
   var calledSetSelf = false;
 
   @override
@@ -433,6 +424,16 @@ class _ReadContext implements AtomContextBase {
 
   @override
   void onDispose(void Function() fn) => disposers.add(fn);
+
+  void dispose() {
+    _disposed = true;
+    if (disposers.isEmpty) return;
+
+    for (final fn in disposers) {
+      fn();
+    }
+    disposers.clear();
+  }
 
   @override
   Value call<Value>(Atom<Value> dep) {
@@ -458,6 +459,11 @@ class _ReadContext implements AtomContextBase {
   @override
   void setSelf(Object? value) {
     calledSetSelf = true;
+
+    if (_disposed) {
+      throw UnsupportedError('can not set a disposed atom');
+    }
+
     store._put(
       atom,
       value,