+++ packages/nucleus/test/atoms/future_test.dart
@@ -14,7 +14,7 @@ void main() {
       final store = Store();
 
       expect(store.read(delayed123), FutureValue.loading());
-      await store.read(delayed123.future);
+      await store.read(delayed123.parent);
       expect(store.read(delayed123), FutureValue.data(123));
     });
 
@@ -40,15 +40,17 @@ void main() {
         return 123;
       });
 
-      expect(store.read(a), FutureValue.loading());
-      await store.read(a.future);
-      expect(store.read(a), FutureValue.data(123));
+      await store.use(a, () async {
+        expect(store.read(a), FutureValue.loading());
+        await store.read(a.parent);
+        expect(store.read(a), FutureValue.data(123));
 
-      store.put(count, 1);
+        store.put(count, 1);
 
-      expect(store.read(a), FutureValue.loading(123));
-      await store.read(a.future);
-      expect(store.read(a), FutureValue.data(123));
+        expect(store.read(a), FutureValue.loading(123));
+        await store.read(a.parent);
+        expect(store.read(a), FutureValue.data(123));
+      });
     });
 
     test('autoDispose works', () async {
@@ -57,11 +59,10 @@ void main() {
       final a = futureAtom((get) async {
         await Future.microtask(() {});
         return 123;
-      })
-        ..autoDispose();
+      });
 
       expect(store.read(a), FutureValue.loading());
-      await store.read(a.future);
+      await store.read(a.parent);
       expect(store.read(a), FutureValue.data(123));
 
       await Future.microtask(() {});